ADD: update config to move entry ids into config.ts;
