+++ sites/grassroots/.gitignore
@@ -30,5 +30,8 @@ yarn-error.log*
 .env.test.local
 .env.production.local
 
+.env.prod
+.env.dev
+
 # vercel
 .vercel
