+++ sites/grassroots/config.ts
@@ -1,0 +1,14 @@
+export const Config = {
+  title: "Grassroots Future",
+  footer:
+    "Grassroots Future is a charitable institution under section 88 of the Inland Revenue Ordinance with No. 91/17131",
+  alt: {
+    logoName: "Grassroots Future logo",
+  },
+
+  contentful: {
+    homePageId: "5xxSi40mvmXQoIJSsmOdeI",
+    aboutPageId: "5Ffkeo9D4oDs3neZrAfq70",
+    featuredPartnersId: "6l4J9blqrSPAEM1de3Ykx8"
+  }
+}
